requirements.txt
absl-py==2.3.1
attrs==25.3.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.59.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
jax==0.6.2
jaxlib==0.6.2
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.44.0
matplotlib==3.10.3
mediapipe==0.10.21
ml_dtypes==0.5.3
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
onnxruntime-gpu==1.19.2
opencv-contrib-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
protobuf==4.25.8
pycparser==2.22
PyMatting==1.1.14
pyparsing==3.2.3
python-dateutil==2.9.0.post0
referencing==0.36.2
rembg==2.0.67
requests==2.32.4
rpds-py==0.26.0
scikit-image==0.25.2
scipy==1.15.3
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.2
sympy==1.14.0
tifffile==2025.5.10
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
